Introduce the Root Circuit to the integration tests (#1471)
### Description

Introduce two new sets of tests in the Integration tests with the root
circuit: one where the root circuit is verified with the MockProver and
the other where a real proof is generated. The input proof to the root
circuit is the output of each subcircuit test we had before.

### Issue Link

Resolve
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1334
Resolve
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1460

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)

### Contents

Issue
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1335
has been resolved by assigning the fixed columns in the ExpTable and
CopyTable in their `load` methods. Previously they were either not
assigned, or assigned dynamically. This was not affecting the
SuperCircuit or the Exp/Copy Circuit because in those cases, the
assignment is done with a different method. Only for SubCircuit tests
where the SubCircuit requires the external table the `load` method is
used.

### Rationale

Testing a Root Circuit (no mater if it's with the MockProver or with a
real proof) requires a real proof of the underlying circuit. This means
that now for each SubCircuit we have 3 tests that require the SubCircuit
proof:
- normal real prover test
- root mock prover test
- root real prover test

I implemented a cache to keep the SubCircuit proofs for each SubCircuit
and block so that we have the possibility of running all the tests in a
single run while reusing the SubCircuit proofs.

### How Has This Been Tested?

I have only tested the SubCircuits (minus SuperCircuit) + Root Circuit
via MockProver.
- The machine I use has 64 GiB of memory, and that's not enough to
generate a SuperCircuit proof
- I have as a pending task to try the Root Circuit real prover

But now we can run these tests under the github CI, so we can defer full
testing after the PR is merged.