Feat/#1474 rw lookup index error (#1482)
### Description

address #1474 and also found #1483 when investigating this issue.

### Issue Link

#1474 

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- The root cause is that there are memory operations in block.get_rws()
but there is only single `copy lookup` expression while copy table
lookup happens in opcode execution. Which makes the indexes of
`assigned_rw_values` and `step.bus_mapping_instance` are inconsistent.
- Add a new function `get_rws_for_check_rw_lookup` specific for
`check_rw_lookup`. This function could bypass memory operations if any
copy table lookup happens, and it makes the indexes of
`assigned_rw_values` and `step.bus_mapping_instance` consistent.

### Rationale

- Add a new function instead of modifying `get_rws()` is because trying
to minimize the impact to our codebase since this is a check function.

---------

Co-authored-by: Ming <hero78119@gmail.com>