Dynamic `CircuitsParameters` (#1445)
### Description

The main goal of this PR is to allow for the dynamic selection of
sub-circuit parameters (`CirctuisParams`) based on the block provided as
input. After this change there are 2 options for selecting circuit
parameters:
  1. Having them fixed before witness generation
2. Leaving them unset as `DynamicCP` and be dynamically generated
together with the witness.


### Issue Link

Close #954

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- Addition of trait `CircuitsParams`, implemented by the `ConcreteCP`
(old `CircuitsParams`) and the new `DynamicCP`
- Addition of generics in all the structs that previously contained
`CircutisParams`: `Block`, `CircuitInputBuilder, `CircuitInputStateRef`
 - Modify unit tests to use dynamic parameters
- Modify variadic test to use default fixed parameters instead of
dynamic parameters

### Rationale


### How Has This Been Tested?

With the existing tests.

### Potential improvements
Sometimes it may be difficult to track which structures contain the
circuit parameters. For clarity, we could remove `CircuitsParams` from
`Block` and create a new structure containing both, something like
`BlockWithParams` or `CircuitInputs`. (suggested by @ed255 )
<hr>

Big thanks to @ed255 for providing the ideas and design on how to solve
this issue !!:partying_face:

---------

Co-authored-by: Eduard S. <eduardsanou@posteo.net>