Hardcode build target in workflow (#1544)
### Description

Remove the matrix strategy in the build job and instead hardcode the
build target. Since we only have 1 target we don't need the matrix
strategy.

This is a small PR aimed to clean up the output of jobs running from
commits merged to main. Our configuration skips jobs that have already
run, and when this happens, the matrix strategy is not resolved and we
see this:

![image](https://github.com/privacy-scaling-explorations/zkevm-circuits/assets/24987329/e90fc7ae-841b-4adb-8956-5b4fa6539fe4)

By removing the matrix strategy we should see a cleaner output.