Fully implement PUSHn (#1527)
### Description
We intend here to support PUSHn with bytecode length inferior to n.

### Issue Link
Issue #633 
Linked to specs issue
https://github.com/privacy-scaling-explorations/zkevm-specs/issues/463

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- Update push.rs to support bytecode of size smaller than n
- Update stop.rs similarly

### Rationale

In pushn, "selectors" are switches to enable or disable lookup arguments
on the bytcode bytes as well as to check whether these are non-null if
needs be. They directly depend on "n" of PUSHn.
We add to these "selectors", "counters" which have similar role but are
a subset of them depending on the bytecode length. We added these extra
switches to be able to easily check the various identities at the risk
of some redundancy.

To support these changes, modifications in stop.py also had to be done.
More precisely, the check "is_out_of_range had to be extended to support
small bytecodes.

### How Has This Been Tested?

Extra tests have been added in push.rs (only).