feat: rust-doc to github pages (#1578)
### Description

This pull request aims to run the cargo doc in CI to generate and
publish the doc to GitHub Pages.

### Issue Link

Solving #1206 

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

The implementation is in _.github/workflows/github-pages.yml_.
The goal is to generate the doc, upload the doc, and deploy the page.

### How Has This Been Tested?

I have tested this in my [forked
repository](https://github.com/ryanycw/zkevm-circuits/actions/workflows/github-pages.yml).
And here is the
[demo](https://ryanycw.github.io/zkevm-circuits/zkevm_circuits/) of what
the result will look like.

---------

Co-authored-by: Chih Cheng Liang <chihchengliang@gmail.com>