fix unconstrained is_warm in sload/sstore (#1596)
### Description

This PR fixes the soundness problem in the implementation of SLOAD and
SSTORE opcodes by adding constraints to the `is_warm` cell. The problem
is detailed in issue #1049.

### Issue Link

#1049

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- Update `bus-mapping/src/evm/opcodes/sload.rs`
- Update `bus-mapping/src/evm/opcodes/sstore.rs`
- Update `zkevm-circuits/src/evm_circuit/execution/sload.rs`
- Update `zkevm-circuits/src/evm_circuit/execution/sstore.rs`

### Rationale

The current implementation of SSTORE and SLOAD didn't constrain the
`is_warm` cell, which could result in any value and lead to soundness
problems. This PR adds constraints to the `is_warm` cell in order to
ensure correctness.

### How Has This Been Tested?

-