fix many bugs of create op (#1657)
### Description

i think this pr can fix many errs of testool report

1. fix bugs of
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1639
2. refactor rw reading using the rw iterator, which is far more friendly
to understand compared to the current manually maintained rw offset way.
3. another refactor here, is move the "init code copy event" from callee
stage to caller stage (before state.push_call). It is more intuitive.
Inside execution, we can think the initcode is "copied from
CopyDataType::Memory to CopyDataType::Bytecode" then is executed. So
doing the copy event inside callee stage is a bit wired.

Btw i have another double on hi-lo (which i am not very familiar with):
is this correct?
https://github.com/privacy-scaling-explorations/zkevm-circuits/blob/09651d40afc59462fd943e9789be8f8b0ccb90d9/zkevm-circuits/src/evm_circuit/execution/create.rs#L545
is_address_collision is boolean, which is not equal to the expr used to
construct the is_zero gadget?

### Issue Link


https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1639

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update