Light client PoC (#1644)
### Description

This PR merges the current milestone's light client PoC into main. Take
into consideration:

- The circuit is almost fished, although there are some small things
that can be to be improved
- There are some changes that maybe needs to be done in the go wrapper
(like sending proofs results in the wrapper call, instead of go code to
retrieve himself) that will requiere changes in the light client. This
is the main cause to currently disable tests since it requieres to
access to a RPC node. In a future changes will be mocked.
- There's still pending some tests passing from a local geth node and
mainnet blocks, when finished, the idea is to reuse some of the code for
the MPT tests, to prove that a set of mainnet block state transitions
can be successfully verified.

### Issue Link

-
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1619
-
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1543

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- A circuit that verifies the state updates requiered to go from one
state root to another
- The witness generator for this circuit, that also, calls internally
the `mpt-witness-generator` wrapper.
- Tests to verify mainnet blocks
- Test to verify a local geth node with `eth_accessListByNumber` RPC
enabled
- A draft of a server to create proofs and serve them. Note that this
can be finished when local geth node tests will be passing.
- Upstreamed MPT golang witness generator changes from miha branch

---------

Co-authored-by: ChihChengLiang <chihchengliang@gmail.com>
Co-authored-by: adria0 <adria0@nowhere>
Co-authored-by: Miha Stopar <stopar.miha@gmail.com>
Co-authored-by: Eduard S <eduardsanou@posteo.net>