Add basic invalid tx support (#1677)
### Description

Resolves
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1611
Upstreamed from https://github.com/taikoxyz/zkevm-circuits/pull/157 with
some additional changes.

Adds an additional execution state for transactions that are invalid and
need to be skipped. This execution state needs to be enabled instead of
the normal BeginTx/EndTx for valid transactions.

Support skipping the following invalid transactions for now:
- nonce too low
- nonce too high
- intrinsic gas too low
- insufficient funds for gas * price + value

More cases will be added in future PRs.

### Issue Link


https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1611

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- [_item_]

### Rationale

[_design decisions and extended information_]

### How Has This Been Tested?

- Added required unit tests
- make test-all passes

---------

Co-authored-by: Cecilia Zhang <zhangyixin319@gmail.com>