chore: set resolver v2 at the workspace level (#1731)
### Description

Cargo defaults to `resolver = "2"` in edition 2021 only for singular
crates. Workspace tho defaults to the old `resolver = "1"` and will
override this setting for all crates in the workspace. [The
guideline](https://doc.rust-lang.org/cargo/reference/workspaces.html#virtual-workspace)
is to manually set the `resolver = "2"` on workspace level.
This was documented recently and since a few releases produces a warning
on building. You don't see it due to the `rust-toolchain` setting.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue) (not really a
bugfix but that's the closest match)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- specifies `resolver = "2"` on workspace level
- removes `resolver = "2"` from light-client-poc/Cargo.toml because it
was ignored anyway

This also fixes this warning:

```
warning: resolver for the non root package will be ignored, specify resolver at the workspace root:
package:   /data/bishop/zkevm-circuits/light-client-poc/Cargo.toml
workspace: /data/bishop/zkevm-circuits/Cargo.toml
```