MPT max nodes (#1745)
### Description

This PR resolves the test failure in
https://github.com/privacy-scaling-explorations/zkevm-circuits/pull/1706.

The parameter `max_nodes` has been added to the MPT to have always the
same fixed column `q_enable`.

The fixed column `q_last` has been removed because it is not necessary -
if some nodes are added to the MPT proof after the account/storage leaf
to enable some non-valid lookups, the proof would fail because of
address/key RLC constraints (the RLC is being built node by node).

### Issue Link


https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1700

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update