MPT StorageDoesNotExist (#1699)
### Description

A couple of `StorageDoesNotExist` issues to be addressed:

1.  Witness generation fails for `getProof` proofs of length 0.

2. For `StorageDoesNotExist`, when the trie is empty, the key
occupies `33 (33 = 161 - 128)` bytes, as in the example below:
```
[227 161 32 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0]
```
Currently, the length for the RLP items is set `valueLen = 34`, so we
don't have space for all 33 bytes.

3. It seems that some constraints of the `StorageDoesNotExist` proof
were lost on the way - currently there are only constraints for the
`wrong` leaf, but the constraints for the case when the node in branch
is `nil` are missing.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

### How Has This Been Tested?

`TestStorageDoesNotExistOnlySProof` in
`mpt-witness-generator/witness/gen_witness_from_infura_blockchain_test.go`.