MPT: Testing & go library fix (#1757)
### Description

This PR:

- Removes the light client (that now is deprecated) and moves its
StateUpdate circuit into bin/mpt-test.
- `bin/mpt-test` is a tool to prove mainnet tests, see the README inside
for more information.
- Is not possible to link to external golang libraries in the same
binary, so this PR also moves golang MPT code inside `/geth-utils`

### Type of change

- [X] New feature (non-breaking change which adds functionality)

### Contents

- Removal if `light-client-poc` folder
- Move `mpt-witness-generator` into `geth-utils`, this includes also
renaming the git action workflow file to geth utils.
- Create new `bin/mpt-folder`
  - `access-lists` contains JSON files of accesslists of mainnet blocks
- `src/circuit` contains a circuit and the witness generator to test
chained MPT proofs. Well, there's some stuff that is not strictly
mandatory to be there - it's taken from light client circuit - but it
can be done later.
- `src/cache.rs` it's a proxy server with a cache file. There's a lot of
RPC calls to execute this tests and this allows to cache them all into a
file, this makes the testing being possible without accessing any RPC
server and speeding-up it.

### How Has This Been Tested?

For the moment, some of the tests are not passing, but the idea is to
integrate into CI when tests passes successfully.

---------

Co-authored-by: Miha Stopar <stopar.miha@gmail.com>
Co-authored-by: adria0 <adria0@nowhere>